migrated code from process_grid_map.py into process_grid.py, removed process_grid_*.py files
#210
