+++ .gitignore
@@ -72,6 +72,7 @@ venv.bak/
 .mypy_cache/
 
 # Others
+.idea/
 /.project
 /.pydevproject
