+++ scripts/process_grid.py
@@ -1,23 +1,32 @@
 import argparse
+from collections import Counter
 from datetime import datetime
 import logging
-import math
 import multiprocessing
-import warnings
+import os
 
-import netCDF4
+from nco import Nco
 import numpy as np
+import xarray as xr
 
-from climate_indices import compute, indices, utils
-from scripts import netcdf_utils
+from climate_indices import compute, indices
 
-# ----------------------------------------------------------------------------------------------------------------------
-# static constants
-_VALID_MIN = -10.0
-_VALID_MAX = 10.0
-_POSSIBLE_MM_UNITS = ["millimeters", "millimeter", "mm"]
-_POSSIBLE_INCH_UNITS = ["inches", "inch"]
-_MM_TO_INCHES_FACTOR = 0.039_370_1
+# the number of worker processes we'll use for process pools
+_NUMBER_OF_WORKER_PROCESSES = multiprocessing.cpu_count() - 1
+
+# shared memory array dictionary keys
+_KEY_ARRAY = "array"
+_KEY_SHAPE = "shape"
+_KEY_LAT = "lat"
+_KEY_RESULT = "result_array"
+_KEY_RESULT_SCPDSI = "result_array_scpdsi"
+_KEY_RESULT_PDSI = "result_array_pdsi"
+_KEY_RESULT_PHDI = "result_array_phdi"
+_KEY_RESULT_PMDI = "result_array_pmdi"
+_KEY_RESULT_ZINDEX = "result_array_zindex"
+
+# global dictionary to contain shared arrays for use by worker processes
+_global_shared_arrays = {}
 
 # ----------------------------------------------------------------------------------------------------------------------
 # set up a basic, global _logger which will write to the console as standard error
@@ -28,948 +37,29 @@ logging.basicConfig(
 )
 _logger = logging.getLogger(__name__)
 
-# ----------------------------------------------------------------------------------------------------------------------
-# multiprocessing locks we'll use to synchronize I/O writes to NetCDF files, one per each output file
-pet_lock = multiprocessing.Lock()
-pdsi_lock = multiprocessing.Lock()
-phdi_lock = multiprocessing.Lock()
-pmdi_lock = multiprocessing.Lock()
-zindex_lock = multiprocessing.Lock()
-scpdsi_lock = multiprocessing.Lock()
-spi_gamma_lock = multiprocessing.Lock()
-spi_pearson_lock = multiprocessing.Lock()
-spei_gamma_lock = multiprocessing.Lock()
-spei_pearson_lock = multiprocessing.Lock()
-pnp_lock = multiprocessing.Lock()
-
-# ignore runtime warnings
-warnings.simplefilter("ignore", Warning)
-
 
 # ----------------------------------------------------------------------------------------------------------------------
-class GridProcessor(object):  # pragma: no cover
-    def __init__(
-        self,
-        index,
-        periodicity,
-        output_file_base,
-        calibration_start_year,
-        calibration_end_year,
-        netcdf_precip,
-        var_name_precip,
-        netcdf_temp=None,
-        var_name_temp=None,
-        netcdf_pet=None,
-        var_name_pet=None,
-        netcdf_awc=None,
-        var_name_awc=None,
-        scales=None,
-    ):
-
-        # assign member values
-        self.output_file_base = output_file_base
-        self.netcdf_precip = netcdf_precip
-        self.netcdf_temp = netcdf_temp
-        self.netcdf_pet = netcdf_pet
-        self.netcdf_awc = netcdf_awc
-        self.var_name_precip = var_name_precip
-        self.var_name_temp = var_name_temp
-        self.var_name_pet = var_name_pet
-        self.var_name_awc = var_name_awc
-        self.scales = scales
-        self.calibration_start_year = calibration_start_year
-        self.calibration_end_year = calibration_end_year
-        self.index = index
-        self.periodicity = periodicity
-
-        # the timestep scale currently (or to next be) computed
-        self.timestep_scale = None
-
-        # determine the file to use for coordinate specs (years range and lat/lon sizes), get relevant units
-        if self.index == "pet":
-
-            # to compute PET we require a temperature input dataset
-            if self.netcdf_temp is not None:
-
-                # a PET file was not provided and we'll compute PET from temperature
-                self.units_temp = netcdf_utils.variable_units(
-                    self.netcdf_temp, self.var_name_temp
-                )
-
-                # use the temperature file as the file that specifies the coordinate specs
-                coordinate_specs_file = self.netcdf_temp
-
-            else:
-                message = (
-                    "A temperature file was not specified, required for PET computation"
-                )
-                _logger.error(message)
-                raise ValueError(message)
-
-        else:
-
-            # all other indices require precipitation
-            if self.netcdf_precip is not None:
-
-                self.units_precip = netcdf_utils.variable_units(
-                    self.netcdf_precip, self.var_name_precip
-                )
-
-            else:
-                message = "A precipitation file was not specified, required for all indices except PET"
-                _logger.error(message)
-                raise ValueError(message)
-
-            # use the precipitation file as the file that specifies the coordinate specs
-            coordinate_specs_file = self.netcdf_precip
-
-            # SPI and PNP only require precipitation
-            if self.index not in ["spi", "pnp"]:
-
-                # PET needs to be either provided or computed from temperature
-                if self.netcdf_pet is not None:
-                    # a PET file was provided
-                    self.units_pet = netcdf_utils.variable_units(
-                        self.netcdf_pet, self.var_name_pet
-                    )
-                elif self.netcdf_temp is not None:
-                    # a PET file was not provided and we'll compute PET from temperature
-                    self.units_temp = netcdf_utils.variable_units(
-                        self.netcdf_temp, self.var_name_temp
-                    )
-                else:
-                    message = (
-                        "Neither a PET nor a temperature file was specified, "
-                        "required for all indices except SPI and PNP"
-                    )
-                    _logger.error(message)
-                    raise ValueError(message)
-
-                # AWC is only required for Palmers
-                if self.index == "palmers":
-
-                    if self.netcdf_awc is not None:
-                        self.units_awc = netcdf_utils.variable_units(
-                            self.netcdf_awc, self.var_name_awc
-                        )
-                        self.fill_value_awc = netcdf_utils.variable_fillvalue(
-                            self.netcdf_awc, self.var_name_awc
-                        )
-                    else:
-                        message = (
-                            "No AWC file was specified, required for Palmers indices"
-                        )
-                        _logger.error(message)
-                        raise ValueError(message)
-
-        # get the coordinate related specs (time range, lat and lon sizes)
-        self.data_start_year, self.data_end_year = netcdf_utils.initial_and_final_years(
-            coordinate_specs_file
-        )
-        self.lat_size, self.lon_size = netcdf_utils.lat_and_lon_sizes(
-            coordinate_specs_file
-        )
-
-        # initialize the NetCDF files used for Palmers output, scaled indices will have
-        # corresponding files initialized at each scale run
-        if self.index == "palmers":
-
-            # place holders for the scaled NetCDFs, these files will be created
-            # and assigned to these variables at each scale's computational iteration
-            self.netcdf_pdsi = self.output_file_base + "_pdsi.nc"
-            self.netcdf_phdi = self.output_file_base + "_phdi.nc"
-            self.netcdf_pmdi = self.output_file_base + "_pmdi.nc"
-            self.netcdf_scpdsi = self.output_file_base + "_scpdsi.nc"
-            self.netcdf_zindex = self.output_file_base + "_zindex.nc"
-
-            netcdf_utils.initialize_netcdf_single_variable_grid(
-                self.netcdf_pdsi,
-                self.netcdf_precip,
-                "pdsi",
-                "Palmer Drought Severity Index",
-                _VALID_MIN,
-                _VALID_MAX,
-            )
-            netcdf_utils.initialize_netcdf_single_variable_grid(
-                self.netcdf_phdi,
-                self.netcdf_precip,
-                "phdi",
-                "Palmer Hydrological Drought Index",
-                _VALID_MIN,
-                _VALID_MAX,
-            )
-            netcdf_utils.initialize_netcdf_single_variable_grid(
-                self.netcdf_pmdi,
-                self.netcdf_precip,
-                "pmdi",
-                "Palmer Modified Drought Index",
-                _VALID_MIN,
-                _VALID_MAX,
-            )
-            netcdf_utils.initialize_netcdf_single_variable_grid(
-                self.netcdf_scpdsi,
-                self.netcdf_precip,
-                "scpdsi",
-                "Self-calibrated Palmer Drought Severity Index",
-                _VALID_MIN,
-                _VALID_MAX,
-            )
-            netcdf_utils.initialize_netcdf_single_variable_grid(
-                self.netcdf_zindex,
-                self.netcdf_precip,
-                "zindex",
-                "Palmer Z-Index",
-                _VALID_MIN,
-                _VALID_MAX,
-            )
-
-        elif self.index in ["spi", "spei", "pnp", "scaled"]:
-
-            # place holders for the scaled NetCDFs, these files will be created as needed
-            # and assigned to these variables at each scale's computational iteration
-            self.netcdf_spi_gamma = ""
-            self.netcdf_spi_pearson = ""
-            self.netcdf_spei_gamma = ""
-            self.netcdf_spei_pearson = ""
-            self.netcdf_pnp = ""
-
-    # ------------------------------------------------------------------------------------------------------------------
-    def _initialize_scaled_netcdfs(self):
-
-        # dictionary of index types to the NetCDF dataset files corresponding to the base index names and
-        # day scales (this is the object we'll build and return from this function)
-        netcdfs = {}
-
-        # make a scale type substring to use within the variable long_name attributes
-        scale_type = str(self.timestep_scale) + "-month scale"
-        if self.periodicity is compute.Periodicity.daily:
-            if self.index in ["spi", "pnp"]:
-                scale_type = str(self.timestep_scale) + "-day scale"
-            else:
-                message = "Incompatible periodicity -- only SPI and PNP are supported for daily time series"
-                _logger.error(message)
-                raise ValueError(message)
-        elif self.periodicity is not compute.Periodicity.monthly:
-            raise ValueError("Unsupported periodicity argument: %s" % self.periodicity)
-
-        # dictionary of index types (ex. 'spi_gamma', 'spei_pearson', etc.) mapped to their corresponding long
-        # variable names, to be used within the respective NetCDFs as variable long_name attributes
-        names_to_longnames = {}
-        if self.index == "spi":
-            names_to_longnames[
-                "spi_gamma"
-            ] = "Standardized Precipitation Index (Gamma distribution), "
-            names_to_longnames[
-                "spi_pearson"
-            ] = "Standardized Precipitation Index (Pearson Type III distribution), "
-        elif self.index == "spei":
-            names_to_longnames["spei_gamma"] = (
-                "Standardized Precipitation Evapotranspiration "
-                "Index (Gamma distribution), "
-            )
-            names_to_longnames["spei_pearson"] = (
-                "Standardized Precipitation Evapotranspiration "
-                "Index (Pearson Type III distribution), "
-            )
-        elif self.index == "pnp":
-            names_to_longnames["pnp"] = "Percentage of Normal Precipitation, "
-        elif self.index == "scaled":
-            names_to_longnames[
-                "spi_gamma"
-            ] = "Standardized Precipitation Index (Gamma distribution), "
-            names_to_longnames[
-                "spi_pearson"
-            ] = "Standardized Precipitation Index (Pearson Type III distribution), "
-            names_to_longnames["spei_gamma"] = (
-                "Standardized Precipitation Evapotranspiration "
-                "Index (Gamma distribution), "
-            )
-            names_to_longnames["spei_pearson"] = (
-                "Standardized Precipitation Evapotranspiration "
-                "Index (Pearson Type III distribution), "
-            )
-            names_to_longnames["pnp"] = "Percentage of Normal Precipitation, "
-        else:
-            raise ValueError("Unsupported index: %s", self.index)
-
-        # add the scale type to the end of each variable's long name
-        names_to_longnames = {
-            k: str(v + scale_type) for (k, v) in names_to_longnames.items()
-        }
-
-        # loop over the indices, creating an output NetCDF dataset for each
-        for index_name, long_name in names_to_longnames.items():
-
-            # use a separate valid min/max for PNP than for the other SP* indices
-            if index_name == "pnp":
-                valid_min = np.float32(-1000.0)
-                valid_max = np.float32(1000.0)
-            else:
-                valid_min = np.float32(-3.09)
-                valid_max = np.float32(3.09)
-
-            # create the variable name from the index and day scale
-            variable_name = index_name + "_{0}".format(
-                str(self.timestep_scale).zfill(2)
-            )
-
-            # create the NetCDF file path from the
-            netcdf_file = self.output_file_base + "_" + variable_name + ".nc"
-
-            # initialize the output NetCDF
-            netcdf_utils.initialize_netcdf_single_variable_grid(
-                netcdf_file,
-                self.netcdf_precip,
-                variable_name,
-                long_name.format(self.timestep_scale),
-                valid_min,
-                valid_max,
-            )
-
-            # add the days scale index's NetCDF to the dictionary for the current index
-            netcdfs[index_name] = netcdf_file
-
-        # assign the NetCDF file paths to the corresponding member variables
-        if self.index == "spi":
-            self.netcdf_spi_gamma = netcdfs["spi_gamma"]
-            self.netcdf_spi_pearson = netcdfs["spi_pearson"]
-        elif self.index == "spei":
-            self.netcdf_spei_gamma = netcdfs["spei_gamma"]
-            self.netcdf_spei_pearson = netcdfs["spei_pearson"]
-        elif self.index == "pnp":
-            self.netcdf_pnp = netcdfs["pnp"]
-        elif self.index == "scaled":
-            self.netcdf_spi_gamma = netcdfs["spi_gamma"]
-            self.netcdf_spi_pearson = netcdfs["spi_pearson"]
-            self.netcdf_spei_gamma = netcdfs["spei_gamma"]
-            self.netcdf_spei_pearson = netcdfs["spei_pearson"]
-            self.netcdf_pnp = netcdfs["pnp"]
-
-    # ------------------------------------------------------------------------------------------------------------------
-    def run(self):
-
-        # the number of worker processes we'll have in our process pool
-        number_of_workers = multiprocessing.cpu_count()  # use 1 here for debugging
-
-        # all index combinations/bundles except SPI and PNP will require PET, so compute it here if required
-        if (self.netcdf_pet is None) and (
-            self.index in ["pet", "spei", "scaled", "palmers"]
-        ):
-            self.netcdf_pet = self.output_file_base + "_pet.nc"
-            netcdf_utils.initialize_netcdf_single_variable_grid(
-                self.netcdf_pet,
-                self.netcdf_temp,
-                "pet",
-                "Potential Evapotranspiration",
-                0.0,
-                10000.0,
-                "millimeters",
-            )
-
-            # create a process Pool for worker processes which will compute indices over an entire latitude slice
-            pool = multiprocessing.Pool(processes=number_of_workers)
-
-            # map the latitude indices as an arguments iterable to the compute function
-            result = pool.map_async(self._process_latitude_pet, range(self.lat_size))
-
-            # get the exception(s) thrown, if any
-            result.get()
-
-            # close the pool and wait on all processes to finish
-            pool.close()
-            pool.join()
-
-        # compute indices other than PET if requested
-        if self.index != "pet":
-
-            if self.index in ["spi", "spei", "pnp", "scaled"]:
-
-                for scale in self.scales:
-                    self.timestep_scale = scale
-
-                    self._initialize_scaled_netcdfs()
-
-                    # create a process Pool for worker processes which will compute indices over a full latitude slice
-                    pool = multiprocessing.Pool(processes=number_of_workers)
-
-                    # map the latitude indices as an arguments iterable to the compute function
-                    result = pool.map_async(
-                        self._process_latitude_scaled, range(self.lat_size)
-                    )
-
-                    # get the exception(s) thrown, if any
-                    result.get()
-
-                    # close the pool and wait on all processes to finish
-                    pool.close()
-                    pool.join()
-
-            elif self.index == "palmers":
-
-                # create a process Pool for worker processes which will compute indices over an entire latitude slice
-                pool = multiprocessing.Pool(processes=number_of_workers)
-
-                # map the latitude indices as an arguments iterable to the compute function
-                result = pool.map_async(
-                    self._process_latitude_palmers, range(self.lat_size)
-                )
-
-                # get the exception(s) thrown, if any
-                result.get()
-
-                # close the pool and wait on all processes to finish
-                pool.close()
-                pool.join()
-
-            else:
-
-                raise ValueError("Unsupported index argument: %s" % self.index)
-
-    # ------------------------------------------------------------------------------------------------------------------
-    def _process_latitude_scaled(self, lat_index):
-        """
-        Processes the relevant scaled indices for a single latitude slice at a single scale.
-
-        :param lat_index: the latitude index of the latitude slice that will be read from NetCDF, computed, and written
-        """
-
-        # open the precipitation NetCDF within a context manager
-        with netCDF4.Dataset(self.netcdf_precip) as dataset_precip:
-
-            # read the latitude slice of input precipitation
-            lat_slice_precip = dataset_precip[self.var_name_precip][
-                lat_index, :, :
-            ]  # assuming (lat, lon, time)
-
-        if self.periodicity is compute.Periodicity.daily:
-
-            # times are daily, transform to all leap year times (i.e. 366 days per year),
-            # so we fill Feb. 29th of each non-leap missing
-            # TODO / FIXME move this up/out of here, should only need to be computed once
-            total_years = self.data_end_year - self.data_start_year + 1
-
-            # allocate an array to hold transformed time series where all years contain 366 days
-            lat_slice_precip_all_leap = np.full(
-                (self.lon_size, total_years * 366), np.NaN
-            )
-
-            # at each longitude we have a time series of values, loop over these longitudes and transform each
-            # corresponding time series to 366 day years representation (fill Feb 29 during non-leap years)
-            # TODO work out how to apply this across the lon axis, to eliminate this loop (np.apply_along_axis()?)
-            for lon_index in range(self.lon_size):
-                # transform the data so it represents all years containing 366 days,
-                # with the Feb. 29th element containing a fill value during non-leap years
-                lat_slice_precip_all_leap[lon_index, :] = utils.to_366day(
-                    lat_slice_precip[lon_index, :], self.data_start_year, total_years
-                )
-
-            # use the all leap daily values as the latitude slice we'll work on
-            lat_slice_precip = lat_slice_precip_all_leap
-
-        # compute PNP if specified
-        if self.index in ["pnp", "scaled"]:
-
-            if self.periodicity is compute.Periodicity.daily:
-                scale_increment = "day"
-            elif self.periodicity is compute.Periodicity.monthly:
-                scale_increment = "month"
-            else:
-                raise ValueError("Invalid periodicity: {}".format(self.periodicity))
-
-            message = "Computing {scale}-{incr} ".format(
-                scale=self.timestep_scale, incr=scale_increment
-            ) + "{index} for latitude index {lat}".format(index="PNP", lat=lat_index)
-            _logger.info(message)
-
-            # compute PNP across all longitudes of the latitude slice
-            lat_slice_pnp = np.apply_along_axis(
-                indices.percentage_of_normal,
-                1,
-                lat_slice_precip,
-                self.timestep_scale,
-                self.data_start_year,
-                self.calibration_start_year,
-                self.calibration_end_year,
-                self.periodicity,
-            )
-
-            if self.periodicity is compute.Periodicity.daily:
-
-                # at each longitude we have a time series of values with a 366 day per year representation
-                # (Feb. 29 during non-leap years is a fill value), loop over these longitudes and transform
-                # each corresponding time series back to a normal Gregorian calendar
-                lat_slice_pnp_gregorian = np.full(
-                    (self.lon_size, lat_slice_precip.shape[1]), np.NaN
-                )
-                for lon_index in range(lat_slice_precip.shape[0]):
-                    # transform the data so it represents mixed leap and non-leap years, i.e. normal Gregorian calendar
-                    lat_slice_pnp_gregorian[lon_index, :] = utils.to_gregorian(
-                        lat_slice_pnp[lon_index, :], self.data_start_year
-                    )
-
-                # use the transformed arrays as the lat slice we'll write to the output NetCDF
-                lat_slice_pnp = lat_slice_pnp_gregorian
-
-            # use relevant variable name
-            pnp_variable_name = "pnp_" + str(self.timestep_scale).zfill(2)
-
-            # open the existing PNP NetCDF file for writing, copy the latitude
-            # slice into the PNP variable at the indexed latitude position
-            pnp_lock.acquire()
-            pnp_dataset = netCDF4.Dataset(self.netcdf_pnp, mode="a")
-            pnp_dataset[pnp_variable_name][
-                lat_index, :, :
-            ] = lat_slice_pnp  # assuming (lat, lon, time) orientation
-            pnp_dataset.sync()
-            pnp_dataset.close()
-            pnp_lock.release()
-
-        # compute SPI if specified
-        if self.index in ["spi", "scaled"]:
-
-            if self.periodicity is compute.Periodicity.daily:
-                scale_increment = "day"
-            elif self.periodicity is compute.Periodicity.monthly:
-                scale_increment = "month"
-            else:
-                raise ValueError("Invalid periodicity: {}".format(self.periodicity))
-
-            _logger.info(
-                "Computing {scale}-{incr} {index} for latitude index {lat}".format(
-                    scale=self.timestep_scale,
-                    incr=scale_increment,
-                    index="SPI",
-                    lat=lat_index,
-                )
-            )
-
-            # compute SPI/Gamma across all longitudes of the latitude slice
-            spi_gamma_lat_slice = np.apply_along_axis(
-                indices.spi,
-                1,
-                lat_slice_precip,
-                self.timestep_scale,
-                indices.Distribution.gamma,
-                self.data_start_year,
-                self.calibration_start_year,
-                self.calibration_end_year,
-                self.periodicity,
-            )
-
-            # compute SPI/Pearson across all longitudes of the latitude slice
-            spi_pearson_lat_slice = np.apply_along_axis(
-                indices.spi,
-                1,
-                lat_slice_precip,
-                self.timestep_scale,
-                indices.Distribution.pearson,
-                self.data_start_year,
-                self.calibration_start_year,
-                self.calibration_end_year,
-                self.periodicity,
-            )
-
-            if self.periodicity is compute.Periodicity.daily:
-
-                # at each longitude we have a time series of values with a 366 day per year representation
-                # (Feb. 29 during non-leap years is a fill value), loop over these longitudes and transform each
-                # corresponding time series back to a normal Gregorian calendar
-                lat_slice_spi_gamma = np.full(
-                    (self.lon_size, lat_slice_precip.shape[1]), np.NaN
-                )
-                lat_slice_spi_pearson = np.full(
-                    (self.lon_size, lat_slice_precip.shape[1]), np.NaN
-                )
-                for lon_index in range(lat_slice_precip.shape[0]):
-                    # transform the data so it represents mixed leap and non-leap years, i.e. normal Gregorian calendar
-                    lat_slice_spi_gamma[lon_index, :] = utils.transform_to_gregorian(
-                        spi_gamma_lat_slice[lon_index, :], self.data_start_year
-                    )
-                    lat_slice_spi_pearson[lon_index, :] = utils.transform_to_gregorian(
-                        spi_pearson_lat_slice[lon_index, :], self.data_start_year
-                    )
-
-                # use these transformed arrays as the lat slices we'll write to the output NetCDF
-                spi_gamma_lat_slice = lat_slice_spi_gamma
-                spi_pearson_lat_slice = lat_slice_spi_pearson
-
-            # use relevant variable names
-            spi_gamma_variable_name = "spi_gamma_" + str(self.timestep_scale).zfill(2)
-            spi_pearson_variable_name = "spi_pearson_" + str(self.timestep_scale).zfill(
-                2
-            )
-
-            # open the existing SPI/Gamma NetCDF file for writing, copy the latitude
-            # slice into the SPI variable at the indexed latitude position
-            spi_gamma_lock.acquire()
-            spi_gamma_dataset = netCDF4.Dataset(self.netcdf_spi_gamma, mode="a")
-            spi_gamma_dataset[spi_gamma_variable_name][
-                lat_index, :, :
-            ] = spi_gamma_lat_slice  # (lat, lon, time)
-            spi_gamma_dataset.sync()
-            spi_gamma_dataset.close()
-            spi_gamma_lock.release()
-
-            # open the existing SPI/Pearson NetCDF file for writing, copy the latitude
-            # slice into the SPI variable at the indexed latitude position
-            spi_pearson_lock.acquire()
-            spi_pearson_dataset = netCDF4.Dataset(self.netcdf_spi_pearson, mode="a")
-            spi_pearson_dataset[spi_pearson_variable_name][
-                lat_index, :, :
-            ] = spi_pearson_lat_slice  # (lat, lon, time)
-            spi_pearson_dataset.sync()
-            spi_pearson_dataset.close()
-            spi_pearson_lock.release()
-
-        # compute SPEI if specified
-        if self.index in ["spei", "scaled"]:
-
-            if self.periodicity is compute.Periodicity.daily:
-                message = "Daily SPEI not yet supported"
-                _logger.error(message)
-                raise ValueError(message)
-            elif self.periodicity is compute.Periodicity.monthly:
-                scale_increment = "month"
-                message = "Computing {scale}-{incr} ".format(
-                    scale=self.timestep_scale, incr=scale_increment
-                ) + "{index} for latitude index {lat}".format(
-                    index="SPEI", lat=lat_index
-                )
-                _logger.info(message)
-
-            # open the PET NetCDF within a context manager (this PET file should be present, either provided initially
-            # as a command line argument to the script or computed from temperature earlier in the processing chain)
-            with netCDF4.Dataset(self.netcdf_pet) as dataset_pet:
-
-                # read the latitude slice of input PET
-                lat_slice_pet = dataset_pet["pet"][
-                    lat_index, :, :
-                ]  # assuming (lat, lon, time) orientation
-
-            # allocate latitude slices for SPEI output
-            spei_gamma_lat_slice = np.full(lat_slice_precip.shape, np.NaN)
-            spei_pearson_lat_slice = np.full(lat_slice_precip.shape, np.NaN)
-
-            # compute SPEI for each longitude from the latitude slice where we have valid inputs
-            for lon_index in range(self.lon_size):
-
-                # get the time series values for this longitude
-                precip_time_series = lat_slice_precip[lon_index, :]
-                pet_time_series = lat_slice_pet[lon_index, :]
-
-                # compute SPEI for the current longitude only if we have valid inputs
-                if (not precip_time_series.mask.all()) and (
-                    not pet_time_series.mask.all()
-                ):
-                    # compute SPEI/Gamma
-                    spei_gamma_lat_slice[lon_index, :] = indices.spei(
-                        precips_mm=precip_time_series,
-                        pet_mm=pet_time_series,
-                        scale=self.timestep_scale,
-                        distribution=indices.Distribution.gamma,
-                        periodicity=self.periodicity,
-                        data_start_year=self.data_start_year,
-                        calibration_year_initial=self.calibration_start_year,
-                        calibration_year_final=self.calibration_end_year,
-                    )
-
-                    # compute SPEI/Pearson
-                    spei_pearson_lat_slice[lon_index, :] = indices.spei(
-                        precips_mm=precip_time_series,
-                        pet_mm=pet_time_series,
-                        scale=self.timestep_scale,
-                        distribution=indices.Distribution.pearson,
-                        periodicity=self.periodicity,
-                        data_start_year=self.data_start_year,
-                        calibration_year_initial=self.calibration_start_year,
-                        calibration_year_final=self.calibration_end_year,
-                    )
-
-            # use relevant variable names
-            spei_gamma_variable_name = "spei_gamma_" + str(self.timestep_scale).zfill(2)
-            spei_pearson_variable_name = "spei_pearson_" + str(
-                self.timestep_scale
-            ).zfill(2)
-
-            # open the existing SPEI/Gamma NetCDF file for writing, copy the latitude slice
-            # into the SPEI variable at the indexed latitude position
-            spei_gamma_lock.acquire()
-            spei_gamma_dataset = netCDF4.Dataset(self.netcdf_spei_gamma, mode="a")
-            spei_gamma_dataset[spei_gamma_variable_name][
-                lat_index, :, :
-            ] = spei_gamma_lat_slice
-            spei_gamma_dataset.sync()
-            spei_gamma_dataset.close()
-            spei_gamma_lock.release()
-
-            # open the existing SPEI/Pearson NetCDF file for writing, copy the latitude slice
-            # into the SPEI variable at the indexed latitude position
-            spei_pearson_lock.acquire()
-            spei_pearson_dataset = netCDF4.Dataset(self.netcdf_spei_pearson, mode="a")
-            spei_pearson_dataset[spei_pearson_variable_name][
-                lat_index, :, :
-            ] = spei_pearson_lat_slice
-            spei_pearson_dataset.sync()
-            spei_pearson_dataset.close()
-            spei_pearson_lock.release()
-
-    # ------------------------------------------------------------------------------------------------------------------
-    def _process_latitude_pet(self, lat_index):
-        """
-        Processes PET for a single latitude slice.
-
-        :param lat_index: the latitude index of the latitude slice that will be read from NetCDF, computed, and written
-        """
-
-        _logger.info(
-            "Computing %s PET for latitude index %s", self.periodicity, lat_index
-        )
-
-        # open the temperature NetCDF within a context manager
-        with netCDF4.Dataset(self.netcdf_temp) as temp_dataset:
-
-            # read the latitude slice of input temperature values
-            temp_lat_slice = temp_dataset[self.var_name_temp][
-                lat_index, :, :
-            ]  # assuming (lat, lon, time) orientation
-
-            # TODO verify that values are in degrees Celsius, if not then convert
-
-            # TODO verify that values are in degrees Celsius, if not then convert
-
-            # get the actual latitude value (assumed to be in degrees north) for the latitude slice
-            latitude_degrees_north = temp_dataset["lat"][lat_index]
-
-            if self.periodicity is compute.Periodicity.daily:
-
-                pass  # placeholder until we work out daily Thornthwaite and/or Hargreaves
-
-            #                 # times are daily, transform to all leap year times (i.e. 366 days per year),
-            #                 # so we fill Feb 29th of each non-leap missing
-            #                 total_years = self.data_end_year - self.data_start_year + 1   # FIXME only needs to be computed once
-            #
-            #                 # allocate an array to hold transformed time series where all years contain 366 days
-            #                 original_days_count = temp_lat_slice.shape[1]
-            #                 temp_lat_slice_all_leap = np.full((self.lon_size, total_years * 366), np.NaN)
-            #
-            #                 # at each longitude we have a time series of values, loop over these longitudes and transform each
-            #                 # corresponding time series to 366 day years representation (fill Feb 29 during non-leap years)
-            #                 for lon_index in range(self.lon_size):  # TODO eliminate loop, work out how to apply across lon axis
-            #
-            #                     # transform the data so it represents all years containing 366 days,
-            #                     # with Feb 29 containing fill value during non-leap years
-            #                     temp_lat_slice_all_leap[lon_index, :] = utils.transform_to_366day(temp_lat_slice[lon_index, :],
-            #                                                                                        self.data_start_year,
-            #                                                                                        total_years)
-            #
-            #                 temp_lat_slice = temp_lat_slice_all_leap
-            #
-            #                 # compute PET across all longitudes of the latitude slice
-            #                 pet_lat_slice = np.apply_along_axis(indices.pet_daily_hargreaves,
-            #                                                     0,
-            #                                                     temp_lat_slice,
-            #                                                     latitude_degrees=latitude_degrees_north,
-            #                                                     data_start_year=self.data_start_year)
-            #
-            #                 # at each longitude we have a time series of values with a 366 day per year representation
-            #                 # (Feb 29 during non-leap years is a fill value), loop over these longitudes and transform
-            #                 # each corresponding time series back to a normal Gregorian calendar
-            #                 lat_slice_pet = np.full((self.lon_size, original_days_count), np.NaN)
-            #                 for lon_index in range(pet_lat_slice.shape[0]):
-            #
-            #                     # transform data so it represents mixed leap and non-leap years, i.e. normal Gregorian calendar
-            #                     lat_slice_pet[lon_index, :] = utils.transform_to_gregorian(pet_lat_slice[lon_index, :],
-            #                                                                                self.data_start_year)
-            #                 pet_lat_slice = lat_slice_pet
-
-            elif self.periodicity is compute.Periodicity.monthly:  # monthly
-
-                # compute PET across all longitudes of the latitude slice
-                pet_lat_slice = np.apply_along_axis(
-                    indices.pet,
-                    1,
-                    temp_lat_slice,
-                    latitude_degrees=latitude_degrees_north,
-                    data_start_year=self.data_start_year,
-                )
-
-            else:
-                raise ValueError("Invalid periodicity: {}".format(self.periodicity))
-
-            # open the existing PET NetCDF file for writing, copy the latitude slice
-            # into the PET variable at the indexed latitude position
-            pet_lock.acquire()
-            pet_dataset = netCDF4.Dataset(self.netcdf_pet, mode="a")
-            # TODO make this more general to allow for other dimension orders, etc.
-            pet_dataset["pet"][
-                lat_index, :, :
-            ] = pet_lat_slice  # this assumes (lat, lon, time)
-            pet_dataset.sync()
-            pet_dataset.close()
-            pet_lock.release()
-
-    # ------------------------------------------------------------------------------------------------------------------
-    def _process_latitude_palmers(self, lat_index):
-        """
-        Perform computation of Palmer indices on a latitude slice, i.e. all lat/lon locations for a single latitude.
-        Each lat/lon will have its corresponding time series used as input, with a corresponding time series output
-        for each index computed. The full latitude slice (all lons for the indexed lat) of index values will be written
-        into the corresponding NetCDF.
-
-        :param lat_index: index of the latitude in the NetCDF, valid range is [0..(total # of divisions - 1)]
-        """
-
-        _logger.info("Computing Palmers for latitude index %s", lat_index)
-
-        # open the input NetCDFs
-        with netCDF4.Dataset(self.netcdf_precip) as dataset_precip, netCDF4.Dataset(
-            self.netcdf_pet
-        ) as dataset_pet, netCDF4.Dataset(self.netcdf_awc) as dataset_awc:
-
-            # read the latitude slice of input precipitation and PET values
-            # assumes (lat, lon, time) orientation
-            lat_slice_precip = dataset_precip[self.var_name_precip][lat_index, :, :]
-            lat_slice_pet = dataset_pet[self.var_name_pet][lat_index, :, :]
-
-            # determine the dimensionality of the AWC dataset, in case there is
-            # a missing time dimension, then get the AWC latitude slice accordingly
-            awc_dims = dataset_awc[self.var_name_awc].dimensions
-            if awc_dims == ("time", "lat", "lon"):
-                awc_lat_slice = dataset_awc[self.var_name_awc][
-                    lat_index, :, 0
-                ].flatten()
-            elif awc_dims == ("lat", "lon"):
-                awc_lat_slice = dataset_awc[self.var_name_awc][lat_index, :].flatten()
-            else:
-                message = (
-                    "Unable to read the available water capacity (AWC) values due to "
-                    + "unsupported variable dimensions: {dims}".format(dims=awc_dims)
-                )
-                _logger.error(message)
-                raise ValueError(message)
-
-        # allocate arrays to contain a latitude slice of Palmer values
-        pdsi_lat_slice = np.full(lat_slice_precip.shape, np.NaN)
-        phdi_lat_slice = np.full(lat_slice_precip.shape, np.NaN)
-        zindex_lat_slice = np.full(lat_slice_precip.shape, np.NaN)
-        scpdsi_lat_slice = np.full(lat_slice_precip.shape, np.NaN)
-        pmdi_lat_slice = np.full(lat_slice_precip.shape, np.NaN)
-
-        # compute Palmer indices for each longitude from the latitude slice where we have valid inputs
-        for lon_index in range(self.lon_size):
-
-            # get the time series values for this longitude
-            precip_time_series = lat_slice_precip[lon_index, :]
-            pet_time_series = lat_slice_pet[lon_index, :]
-            awc = awc_lat_slice[lon_index]
-
-            # compute Palmer indices only if we have valid inputs
-            if (
-                utils.is_data_valid(precip_time_series)
-                and utils.is_data_valid(pet_time_series)
-                and awc is not np.ma.masked
-                and not math.isnan(awc)
-                and not math.isclose(awc, self.fill_value_awc)
-            ):
-
-                # put precipitation and PET into inches, if not already
-                if self.units_precip in _POSSIBLE_MM_UNITS:
-                    precip_time_series = precip_time_series * _MM_TO_INCHES_FACTOR
-                if self.units_pet in _POSSIBLE_MM_UNITS:
-                    pet_time_series = pet_time_series * _MM_TO_INCHES_FACTOR
-
-                # compute Palmer indices
-                scpdsi, pdsi, phdi, pmdi, zindex = indices.scpdsi(
-                    precip_time_series,
-                    pet_time_series,
-                    awc,
-                    self.data_start_year,
-                    self.calibration_start_year,
-                    self.calibration_end_year,
-                )
-
-                # add the values into the slice, first clipping all values to the valid range
-                scpdsi_lat_slice[lon_index, :] = np.clip(scpdsi, _VALID_MIN, _VALID_MAX)
-                pdsi_lat_slice[lon_index, :] = np.clip(pdsi, _VALID_MIN, _VALID_MAX)
-                phdi_lat_slice[lon_index, :] = np.clip(phdi, _VALID_MIN, _VALID_MAX)
-                pmdi_lat_slice[lon_index, :] = np.clip(pmdi, _VALID_MIN, _VALID_MAX)
-                zindex_lat_slice[lon_index, :] = zindex
-
-        # open the existing PDSI NetCDF file for writing, copy the latitude slice
-        # into the PET variable at the indexed latitude position
-        pdsi_lock.acquire()
-        pdsi_dataset = netCDF4.Dataset(self.netcdf_pdsi, mode="a")
-        pdsi_dataset["pdsi"][lat_index, :, :] = pdsi_lat_slice
-        pdsi_dataset.sync()
-        pdsi_dataset.close()
-        pdsi_lock.release()
-
-        # open the existing PHDI NetCDF file for writing, copy the latitude slice
-        # into the PET variable at the indexed latitude position
-        phdi_lock.acquire()
-        phdi_dataset = netCDF4.Dataset(self.netcdf_phdi, mode="a")
-        phdi_dataset["phdi"][lat_index, :, :] = phdi_lat_slice
-        phdi_dataset.sync()
-        phdi_dataset.close()
-        phdi_lock.release()
-
-        # open the existing Z-Index NetCDF file for writing, copy the latitude slice
-        # into the Z-Index variable at the indexed latitude position
-        zindex_lock.acquire()
-        zindex_dataset = netCDF4.Dataset(self.netcdf_zindex, mode="a")
-        zindex_dataset["zindex"][lat_index, :, :] = zindex_lat_slice
-        zindex_dataset.sync()
-        zindex_dataset.close()
-        zindex_lock.release()
-
-        # open the existing SCPDSI NetCDF file for writing, copy the latitude slice
-        # into the scPDSI variable at the indexed latitude position
-        scpdsi_lock.acquire()
-        scpdsi_dataset = netCDF4.Dataset(self.netcdf_scpdsi, mode="a")
-        scpdsi_dataset["scpdsi"][lat_index, :, :] = scpdsi_lat_slice
-        scpdsi_dataset.sync()
-        scpdsi_dataset.close()
-        scpdsi_lock.release()
-
-        # open the existing PMDI NetCDF file for writing, copy the latitude slice
-        # into the PMDI variable at the indexed latitude position
-        pmdi_lock.acquire()
-        pmdi_dataset = netCDF4.Dataset(self.netcdf_pmdi, mode="a")
-        pmdi_dataset["pmdi"][lat_index, :, :] = pmdi_lat_slice
-        pmdi_dataset.sync()
-        pmdi_dataset.close()
-        pmdi_lock.release()
+def init_worker(arrays_and_shapes):
+    """
+    Initialization function that assigns named arrays into the global variable.
+
+    :param arrays_and_shapes: dictionary containing variable names as keys
+        and two-element dictionaries containing RawArrays and associated shapes
+        (i.e. each value of the dictionary is itself a dictionary with one key "array"
+        and another key _KEY_SHAPE)
+    :return:
+    """
+
+    global _global_shared_arrays
+    _global_shared_arrays = arrays_and_shapes
 
 
 # ----------------------------------------------------------------------------------------------------------------------
-def _validate_arguments(
-    index,
-    periodicity,
-    netcdf_precip,
-    var_name_precip,
-    netcdf_temp=None,
-    var_name_temp=None,
-    netcdf_pet=None,
-    var_name_pet=None,
-    netcdf_awc=None,
-    var_name_awc=None,
-    scales=None,
-):
+def _validate_args(args):
     """
     Validate the processing settings to confirm that proper argument combinations have been provided.
-    
-    :param index:
-    :param periodicity:
-    :param netcdf_precip:
-    :param var_name_precip:
-    :param netcdf_temp:
-    :param var_name_temp:
-    :param netcdf_pet:
-    :param var_name_pet:
-    :param netcdf_awc:
-    :param var_name_awc:
-    :param scales:
-    :return:
+
+    :param args: an arguments object of the type returned by argparse.ArgumentParser.parse_args()
     :raise ValueError: if one or more of the command line arguments is invalid
     """
 
@@ -977,35 +67,35 @@ def _validate_arguments(
     expected_dimensions = [("lat", "lon", "time"), ("time", "lat", "lon")]
 
     # all indices except PET require a precipitation file
-    if index != "pet":
+    if args.index != "pet":
 
         # make sure a precipitation file was specified
-        if netcdf_precip is None:
+        if args.netcdf_precip is None:
             msg = "Missing the required precipitation file"
             _logger.error(msg)
             raise ValueError(msg)
 
         # make sure a precipitation variable name was specified
-        if var_name_precip is None:
+        if args.var_name_precip is None:
             message = "Missing precipitation variable name"
             _logger.error(message)
             raise ValueError(message)
 
         # validate the precipitation file itself
-        with netCDF4.Dataset(netcdf_precip) as dataset_precip:
+        with xr.open_dataset(args.netcdf_precip) as dataset_precip:
 
             # make sure we have a valid precipitation variable name
-            if var_name_precip not in dataset_precip.variables:
+            if args.var_name_precip not in dataset_precip.variables:
                 message = "Invalid precipitation variable name: '{var}' ".format(
-                    var=var_name_precip
+                    var=args.var_name_precip
                 ) + "does not exist in precipitation file '{file}'".format(
-                    file=netcdf_precip
+                    file=args.netcdf_precip
                 )
                 _logger.error(message)
                 raise ValueError(message)
 
             # verify that the precipitation variable's dimensions are in the expected order
-            dimensions = dataset_precip.variables[var_name_precip].dimensions
+            dimensions = dataset_precip[args.var_name_precip].dims
             if dimensions not in expected_dimensions:
                 message = "Invalid dimensions of the precipitation variable: {dims}, ".format(
                     dims=dimensions
@@ -1016,54 +106,58 @@ def _validate_arguments(
                 raise ValueError(message)
 
             # get the sizes of the latitude and longitude coordinate variables
-            lats_precip = dataset_precip.variables["lat"][:]
-            lons_precip = dataset_precip.variables["lon"][:]
-            times_precip = dataset_precip.variables["time"][:]
+            lats_precip = dataset_precip["lat"].values[:]
+            lons_precip = dataset_precip["lon"].values[:]
+            times_precip = dataset_precip["time"].values[:]
 
     else:
 
         # PET requires a temperature file
-        if netcdf_temp is None:
+        if args.netcdf_temp is None:
             msg = "Missing the required temperature file argument"
             _logger.error(msg)
             raise ValueError(msg)
 
         # don't allow a daily periodicity (yet, this will be possible once we have Hargreaves or a daily Thornthwaite)
-        if periodicity is not compute.Periodicity.monthly:
+        if args.periodicity is not compute.Periodicity.monthly:
             msg = (
                 "Invalid periodicity argument for PET: "
-                + "'{period}' -- only monthly is supported".format(period=periodicity)
+                + "'{period}' -- only monthly is supported".format(
+                    period=args.periodicity
+                )
             )
             _logger.error(msg)
             raise ValueError(msg)
 
     # SPEI and Palmers require either a PET file or a temperature file in order to compute PET
-    if index in ["spei", "scaled", "palmers"]:
+    if args.index in ["spei", "scaled", "palmers"]:
 
-        if netcdf_temp is None:
+        if args.netcdf_temp is None:
 
-            if netcdf_pet is None:
+            if args.netcdf_pet is None:
                 msg = "Missing the required temperature or PET files, neither were provided"
                 _logger.error(msg)
                 raise ValueError(msg)
 
             # validate the PET file
-            with netCDF4.Dataset(netcdf_pet) as dataset_pet:
+            with xr.open_dataset(args.netcdf_pet) as dataset_pet:
 
                 # make sure we have a valid PET variable name
-                if var_name_pet is None:
+                if args.var_name_pet is None:
                     message = "Missing PET variable name"
                     _logger.error(message)
                     raise ValueError(message)
-                elif var_name_pet not in dataset_pet.variables:
+                elif args.var_name_pet not in dataset_pet.variables:
                     message = "Invalid PET variable name: '{var_name}' ".format(
-                        var_name=var_name_pet
-                    ) + "does not exist in PET file '{file}'".format(file=netcdf_pet)
+                        var_name=args.var_name_pet
+                    ) + "does not exist in PET file '{file}'".format(
+                        file=args.netcdf_pet
+                    )
                     _logger.error(message)
                     raise ValueError(message)
 
                 # verify that the PET variable's dimensions are in the expected order
-                dimensions = dataset_pet.variables[var_name_pet].dimensions
+                dimensions = dataset_pet[args.var_name_pet].dims
                 if dimensions not in expected_dimensions:
                     message = "Invalid dimensions of the PET variable: {dims}, ".format(
                         dims=dimensions
@@ -1074,24 +168,24 @@ def _validate_arguments(
                     raise ValueError(message)
 
                 # verify that the coordinate variables match with those of the precipitation dataset
-                if not np.array_equal(lats_precip, dataset_pet.variables["lat"][:]):
+                if not np.array_equal(lats_precip, dataset_pet["lat"][:]):
                     message = (
                         "Precipitation and PET variables contain non-matching latitudes"
                     )
                     _logger.error(message)
                     raise ValueError(message)
-                elif not np.array_equal(lons_precip, dataset_pet.variables["lon"][:]):
+                elif not np.array_equal(lons_precip, dataset_pet["lon"][:]):
                     message = "Precipitation and PET variables contain non-matching longitudes"
                     _logger.error(message)
                     raise ValueError(message)
-                elif not np.array_equal(times_precip, dataset_pet.variables["time"][:]):
+                elif not np.array_equal(times_precip, dataset_pet["time"][:]):
                     message = (
                         "Precipitation and PET variables contain non-matching times"
                     )
                     _logger.error(message)
                     raise ValueError(message)
 
-        elif netcdf_pet is not None:
+        elif args.netcdf_pet is not None:
 
             # we can't have both temperature and PET files specified, no way to determine which to use
             msg = "Both temperature and PET files were specified, only one of these should be provided"
@@ -1101,24 +195,24 @@ def _validate_arguments(
         else:
 
             # validate the temperature file
-            with netCDF4.Dataset(netcdf_temp) as dataset_temp:
+            with xr.open_dataset(args.netcdf_temp) as dataset_temp:
 
                 # make sure we have a valid temperature variable name
-                if var_name_temp is None:
+                if args.var_name_temp is None:
                     message = "Missing temperature variable name"
                     _logger.error(message)
                     raise ValueError(message)
-                elif var_name_temp not in dataset_temp.variables:
+                elif args.var_name_temp not in dataset_temp.variables:
                     message = "Invalid temperature variable name: '{var}' does ".format(
-                        var=var_name_temp
+                        var=args.var_name_temp
                     ) + "not exist in temperature file '{file}'".format(
-                        file=netcdf_temp
+                        file=args.netcdf_temp
                     )
                     _logger.error(message)
                     raise ValueError(message)
 
                 # verify that the temperature variable's dimensions are in the expected order
-                dimensions = dataset_temp.variables[var_name_temp].dimensions
+                dimensions = dataset_temp[args.var_name_temp].dims
                 if dimensions not in expected_dimensions:
                     message = "Invalid dimensions of the temperature variable: {dims}, ".format(
                         dims=dimensions
@@ -1129,51 +223,50 @@ def _validate_arguments(
                     raise ValueError(message)
 
                 # verify that the coordinate variables match with those of the precipitation dataset
-                if not np.array_equal(lats_precip, dataset_temp.variables["lat"][:]):
+                if not np.array_equal(lats_precip, dataset_temp["lat"][:]):
                     message = "Precipitation and temperature variables contain non-matching latitudes"
                     _logger.error(message)
                     raise ValueError(message)
-                elif not np.array_equal(lons_precip, dataset_temp.variables["lon"][:]):
+                elif not np.array_equal(lons_precip, dataset_temp["lon"][:]):
                     message = "Precipitation and temperature variables contain non-matching longitudes"
                     _logger.error(message)
                     raise ValueError(message)
-                elif not np.array_equal(
-                    times_precip, dataset_temp.variables["time"][:]
-                ):
+                elif not np.array_equal(times_precip, dataset_temp["time"][:]):
                     message = "Precipitation and temperature variables contain non-matching times"
                     _logger.error(message)
                     raise ValueError(message)
 
         # Palmers requires an available water capacity file
-        if index in ["palmers"]:
+        if args.index in ["palmers"]:
 
-            if netcdf_awc is None:
+            if args.netcdf_awc is None:
+
                 msg = "Missing the required available water capacity file"
                 _logger.error(msg)
                 raise ValueError(msg)
 
             # validate the AWC file
-            with netCDF4.Dataset(netcdf_awc) as dataset_awc:
+            with xr.open_dataset(args.netcdf_awc) as dataset_awc:
 
                 # make sure we have a valid PET variable name
-                if var_name_awc is None:
+                if args.var_name_awc is None:
                     message = "Missing the AWC variable name"
                     _logger.error(message)
                     raise ValueError(message)
-                elif var_name_awc not in dataset_awc.variables:
+                elif args.var_name_awc not in dataset_awc.variables:
                     message = "Invalid AWC variable name: '{var}' does not exist ".format(
-                        var=var_name_awc
+                        var=args.var_name_awc
                     ) + "in AWC file '{file}'".format(
-                        file=netcdf_awc
+                        file=args.netcdf_awc
                     )
                     _logger.error(message)
                     raise ValueError(message)
 
                 # verify that the AWC variable's dimensions are in the expected order
-                dimensions = dataset_awc.variables[var_name_awc].dimensions
-                if (dimensions != ("lat", "lon")) and (
-                    dimensions != expected_dimensions
-                ):
+                dimensions = dataset_awc[args.var_name_awc].dims
+                if len(dimensions) == 2:
+                    expected_dimensions = [("lat", "lon")]
+                if dimensions not in expected_dimensions:
                     message = "Invalid dimensions of the AWC variable: {dims}, ".format(
                         dims=dimensions
                     ) + "(expected names and order: {dims})".format(
@@ -1182,21 +275,22 @@ def _validate_arguments(
                     _logger.error(message)
                     raise ValueError(message)
 
-                # verify that the lat and lon coordinate variables match with those of the precipitation dataset
-                if not np.array_equal(lats_precip, dataset_awc.variables["lat"][:]):
+                # verify that the lat and lon coordinate variable values
+                # match with those of the precipitation dataset
+                if not np.array_equal(lats_precip, dataset_awc["lat"][:]):
                     message = (
                         "Precipitation and AWC variables contain non-matching latitudes"
                     )
                     _logger.error(message)
                     raise ValueError(message)
-                elif not np.array_equal(lons_precip, dataset_awc.variables["lon"][:]):
+                elif not np.array_equal(lons_precip, dataset_awc["lon"][:]):
                     message = "Precipitation and AWC variables contain non-matching longitudes"
                     _logger.error(message)
                     raise ValueError(message)
 
-    if index in ["spi", "spei", "scaled", "pnp"]:
+    if args.index in ["spi", "spei", "scaled", "pnp"]:
 
-        if scales is None:
+        if args.scales is None:
             message = (
                 "Scaled indices (SPI, SPEI, and/or PNP) specified without including "
                 + "one or more time scales (missing --scales argument)"
@@ -1204,79 +298,952 @@ def _validate_arguments(
             _logger.error(message)
             raise ValueError(message)
 
-        if any(n < 0 for n in scales):
+        if any(n < 0 for n in args.scales):
             message = "One or more negative scale specified within --scales argument"
             _logger.error(message)
             raise ValueError(message)
 
 
 # ----------------------------------------------------------------------------------------------------------------------
-def process_grid(
-    index,
-    periodicity,
-    output_file_base,
-    calibration_start_year,
-    calibration_end_year,
-    netcdf_precip,
-    var_name_precip,
-    netcdf_temp=None,
-    var_name_temp=None,
-    netcdf_pet=None,
-    var_name_pet=None,
-    netcdf_awc=None,
-    var_name_awc=None,
-    scales=None,
-):
-    # validate the arguments
-    _validate_arguments(
-        index,
-        periodicity,
-        netcdf_precip,
-        var_name_precip,
-        netcdf_temp,
-        var_name_temp,
-        netcdf_pet,
-        var_name_pet,
-        netcdf_awc,
-        var_name_awc,
-        scales,
+def _get_scale_increment(args_dict):
+
+    if args_dict["periodicity"] == compute.Periodicity.daily:
+        scale_increment = "day"
+    elif args_dict["periodicity"] == compute.Periodicity.monthly:
+        scale_increment = "month"
+    else:
+        raise ValueError(
+            "Invalid periodicity argument: {}".format(args_dict["periodicity"])
+        )
+
+    return scale_increment
+
+
+# ----------------------------------------------------------------------------------------------------------------------
+def _log_status(args_dict):
+
+    # get the scale increment for use in later log messages
+    if "scale" in args_dict:
+
+        if "distribution" in args_dict:
+
+            _logger.info(
+                "Computing {scale}-{incr} {index}/{dist}".format(
+                    scale=args_dict["scale"],
+                    incr=_get_scale_increment(args_dict),
+                    index=args_dict["index"].upper(),
+                    dist=args_dict["distribution"].value.capitalize(),
+                )
+            )
+
+        else:
+
+            _logger.info(
+                "Computing {scale}-{incr} {index}".format(
+                    scale=args_dict["scale"],
+                    incr=_get_scale_increment(args_dict),
+                    index=args_dict["index"].upper(),
+                )
+            )
+
+    else:
+
+        _logger.info("Computing {index}".format(index=args_dict["index"].upper()))
+
+    return True
+
+
+# ----------------------------------------------------------------------------------------------------------------------
+def _build_arguments(keyword_args):
+    """
+    Builds a dictionary of function arguments appropriate to the index to be computed.
+
+    :param dict keyword_args:
+    :return: dictionary of arguments keyed with names expected by the corresponding
+        index computation function
+    """
+
+    function_arguments = {"data_start_year": keyword_args["data_start_year"]}
+
+    if keyword_args["index"] in ["spi", "spei"]:
+        function_arguments["scale"] = keyword_args["scale"]
+        function_arguments["distribution"] = keyword_args["distribution"]
+        function_arguments["calibration_year_initial"] = keyword_args[
+            "calibration_start_year"
+        ]
+        function_arguments["calibration_year_final"] = keyword_args[
+            "calibration_end_year"
+        ]
+        function_arguments["periodicity"] = keyword_args["periodicity"]
+
+    elif keyword_args["index"] == "pnp":
+        function_arguments["scale"] = keyword_args["scale"]
+        function_arguments["calibration_start_year"] = keyword_args[
+            "calibration_start_year"
+        ]
+        function_arguments["calibration_end_year"] = keyword_args[
+            "calibration_end_year"
+        ]
+        function_arguments["periodicity"] = keyword_args["periodicity"]
+
+    elif keyword_args["index"] == "palmers":
+        function_arguments["calibration_start_year"] = keyword_args[
+            "calibration_start_year"
+        ]
+        function_arguments["calibration_end_year"] = keyword_args[
+            "calibration_end_year"
+        ]
+
+    elif keyword_args["index"] != "pet":
+        raise ValueError(
+            "Index {index} not yet supported.".format(index=keyword_args["index"])
+        )
+
+    return function_arguments
+
+
+# ----------------------------------------------------------------------------------------------------------------------
+def _get_variable_attributes(args_dict):
+
+    if args_dict["index"] == "spi":
+
+        long_name = "Standardized Precipitation Index ({dist} distribution), ".format(
+            dist=args_dict["distribution"].value.capitalize()
+        ) + "{scale}-{increment}".format(
+            scale=args_dict["scale"], increment=_get_scale_increment(args_dict)
+        )
+        attrs = {"long_name": long_name, "valid_min": -3.09, "valid_max": 3.09}
+        var_name = (
+            "spi_"
+            + args_dict["distribution"].value
+            + "_"
+            + str(args_dict["scale"]).zfill(2)
+        )
+
+    elif args_dict["index"] == "spei":
+
+        long_name = "Standardized Precipitation Evapotranspiration Index ({dist} distribution), ".format(
+            dist=args_dict["distribution"].value.capitalize()
+        ) + "{scale}-{increment}".format(
+            scale=args_dict["scale"], increment=_get_scale_increment(args_dict)
+        )
+        attrs = {"long_name": long_name, "valid_min": -3.09, "valid_max": 3.09}
+        var_name = (
+            "spei_"
+            + args_dict["distribution"].value
+            + "_"
+            + str(args_dict["scale"]).zfill(2)
+        )
+
+    elif args_dict["index"] == "pnp":
+
+        long_name = (
+            "Percentage of Normal Precipitation, "
+            + "{scale}-{increment}".format(
+                scale=args_dict["scale"], increment=_get_scale_increment(args_dict)
+            )
+        )
+        attrs = {"long_name": long_name, "valid_min": -1000.0, "valid_max": 1000.0}
+        var_name = "pnp_" + str(args_dict["scale"]).zfill(2)
+
+    elif args_dict["index"] == "pet":
+
+        long_name = "Potential Evapotranspiration (Thornthwaite)"
+        attrs = {
+            "long_name": long_name,
+            "valid_min": 0.0,
+            "valid_max": 10000.0,
+            "units": "millimeters",
+        }
+        var_name = "pet_thornthwaite"
+
+    else:
+
+        raise ValueError("Unsupported index: {index}".format(index=args_dict["index"]))
+
+    return var_name, attrs
+
+
+# ----------------------------------------------------------------------------------------------------------------------
+def _compute_write_index(keyword_arguments):
+    """
+    Computes a climate index and writes the result into a corresponding NetCDF.
+
+    :param keyword_arguments:
+    :return:
+    """
+
+    _log_status(keyword_arguments)
+
+    # open the NetCDF files as an xarray DataSet object
+    files = []
+    if "netcdf_precip" in keyword_arguments:
+        files.append(keyword_arguments["netcdf_precip"])
+    if "netcdf_temp" in keyword_arguments:
+        files.append(keyword_arguments["netcdf_temp"])
+    if "netcdf_pet" in keyword_arguments:
+        files.append(keyword_arguments["netcdf_pet"])
+    dataset = xr.open_mfdataset(files, chunks={"lat": -1, "lon": -1})
+
+    # trim out all data variables from the dataset except the ones we'll need
+    input_var_names = []
+    if "var_name_precip" in keyword_arguments:
+        input_var_names.append(keyword_arguments["var_name_precip"])
+    if "var_name_temp" in keyword_arguments:
+        input_var_names.append(keyword_arguments["var_name_temp"])
+    if "var_name_pet" in keyword_arguments:
+        input_var_names.append(keyword_arguments["var_name_pet"])
+    for var in dataset.data_vars:
+        if var not in input_var_names:
+            dataset = dataset.drop(var)
+
+    # get the initial year of the data
+    data_start_year = int(str(dataset["time"].values[0])[0:4])
+    keyword_arguments["data_start_year"] = data_start_year
+
+    # the shape of output variables is assumed to match that of the input,
+    # so use either precipitation or temperature variable's shape
+    if "var_name_precip" in keyword_arguments:
+        output_shape = dataset[keyword_arguments["var_name_precip"]].shape
+        output_dims = dataset[keyword_arguments["var_name_precip"]].dims
+    elif "var_name_temp" in keyword_arguments:
+        output_shape = dataset[keyword_arguments["var_name_temp"]].shape
+        output_dims = dataset[keyword_arguments["var_name_temp"]].dims
+    else:
+        raise ValueError(
+            "Unable to determine output shape, no precipitation "
+            "or temperature variable name was specified."
+        )
+
+    # get the data arrays we'll use later in the index computations
+    global _global_shared_arrays
+    expected_dims_3d = (("lat", "lon", "time"), ("lon", "lat", "time"))
+    expected_dims_2d = (("lat", "lon"), ("lon", "lat"))
+    for var_name in input_var_names:
+
+        # confirm that the dimensions of the data array are valid
+        dims = dataset[var_name].dims
+        if len(dims) == 3:
+            if dims not in expected_dims_3d:
+                message = "Invalid dimensions for variable '{var_name}': {dims}".format(
+                    var_name=var_name, dims=dims
+                )
+                _logger.error(message)
+                raise ValueError(message)
+        elif len(dims) == 2:
+            if dims not in expected_dims_2d:
+                message = "Invalid dimensions for variable '{var_name}': {dims}".format(
+                    var_name=var_name, dims=dims
+                )
+                _logger.error(message)
+                raise ValueError(message)
+
+        # create a shared memory array, wrap it as a numpy array and copy
+        # copy the data (values) from this variable's DataArray
+        shared_array = multiprocessing.Array("d", int(np.prod(dataset[var_name].shape)))
+        shared_array_np = np.frombuffer(shared_array.get_obj()).reshape(
+            dataset[var_name].shape
+        )
+        np.copyto(shared_array_np, dataset[var_name].values)
+
+        # add to the dictionary of arrays
+        _global_shared_arrays[var_name] = {
+            _KEY_ARRAY: shared_array,
+            _KEY_SHAPE: dataset[var_name].shape,
+        }
+
+        # drop the variable from the dataset (we're assuming this frees the memory)
+        dataset = dataset.drop(var_name)
+
+    # build an arguments dictionary appropriate to the index we'll compute
+    args = _build_arguments(keyword_arguments)
+
+    # add output variable arrays into the shared memory arrays dictionary
+    if keyword_arguments["index"] == "palmers":
+
+        # read AWC data into shared memory array
+        if ("netcdf_awc" not in keyword_arguments) or (
+            "var_name_awc" not in keyword_arguments
+        ):
+            raise ValueError("Missing the AWC file and/or variable name argument(s)")
+
+        awc_dataset = xr.open_dataset(keyword_arguments["netcdf_awc"])
+
+        # create a shared memory array, wrap it as a numpy array and copy
+        # copy the data (values) from this variable's DataArray
+        var_name = keyword_arguments["var_name_awc"]
+        shared_array = multiprocessing.Array(
+            "d", int(np.prod(awc_dataset[var_name].shape))
+        )
+        shared_array_np = np.frombuffer(shared_array.get_obj()).reshape(
+            awc_dataset[var_name].shape
+        )
+        np.copyto(shared_array_np, awc_dataset[var_name].values)
+
+        # add to the dictionary of arrays
+        _global_shared_arrays[var_name] = {
+            _KEY_ARRAY: shared_array,
+            _KEY_SHAPE: awc_dataset[var_name].shape,
+        }
+
+        # add shared memory arrays for computed Palmers to the dictionary of shared arrays
+        if _KEY_RESULT_SCPDSI not in _global_shared_arrays:
+            _global_shared_arrays[_KEY_RESULT_SCPDSI] = {
+                _KEY_ARRAY: multiprocessing.Array("d", int(np.prod(output_shape))),
+                _KEY_SHAPE: output_shape,
+            }
+        if _KEY_RESULT_PDSI not in _global_shared_arrays:
+            _global_shared_arrays[_KEY_RESULT_PDSI] = {
+                _KEY_ARRAY: multiprocessing.Array("d", int(np.prod(output_shape))),
+                _KEY_SHAPE: output_shape,
+            }
+        if _KEY_RESULT_PHDI not in _global_shared_arrays:
+            _global_shared_arrays[_KEY_RESULT_PHDI] = {
+                _KEY_ARRAY: multiprocessing.Array("d", int(np.prod(output_shape))),
+                _KEY_SHAPE: output_shape,
+            }
+        if _KEY_RESULT_PMDI not in _global_shared_arrays:
+            _global_shared_arrays[_KEY_RESULT_PMDI] = {
+                _KEY_ARRAY: multiprocessing.Array("d", int(np.prod(output_shape))),
+                _KEY_SHAPE: output_shape,
+            }
+        if _KEY_RESULT_ZINDEX not in _global_shared_arrays:
+            _global_shared_arrays[_KEY_RESULT_ZINDEX] = {
+                _KEY_ARRAY: multiprocessing.Array("d", int(np.prod(output_shape))),
+                _KEY_SHAPE: output_shape,
+            }
+
+        # apply the Palmers function along the time axis (axis=2)
+        _parallel_process(
+            keyword_arguments["index"],
+            _global_shared_arrays,
+            {
+                "var_name_precip": keyword_arguments["var_name_precip"],
+                "var_name_pet": keyword_arguments["var_name_pet"],
+                "var_name_awc": keyword_arguments["var_name_awc"],
+            },
+            _KEY_RESULT_SCPDSI,
+            args,
+        )
+
+        # get the computed SCPDSI data as an array of float32 values
+        array = _global_shared_arrays[_KEY_RESULT_SCPDSI][_KEY_ARRAY]
+        shape = _global_shared_arrays[_KEY_RESULT_SCPDSI][_KEY_SHAPE]
+        scpdsi = np.frombuffer(array.get_obj()).reshape(shape).astype(np.float32)
+
+        # get the computedPDSI data as an array of float32 values
+        array = _global_shared_arrays[_KEY_RESULT_PDSI][_KEY_ARRAY]
+        shape = _global_shared_arrays[_KEY_RESULT_PDSI][_KEY_SHAPE]
+        pdsi = np.frombuffer(array.get_obj()).reshape(shape).astype(np.float32)
+
+        # get the computed PHDI data as an array of float32 values
+        array = _global_shared_arrays[_KEY_RESULT_PHDI][_KEY_ARRAY]
+        shape = _global_shared_arrays[_KEY_RESULT_PHDI][_KEY_SHAPE]
+        phdi = np.frombuffer(array.get_obj()).reshape(shape).astype(np.float32)
+
+        # get the computed PMDI data as an array of float32 values
+        array = _global_shared_arrays[_KEY_RESULT_PMDI][_KEY_ARRAY]
+        shape = _global_shared_arrays[_KEY_RESULT_PMDI][_KEY_SHAPE]
+        pmdi = np.frombuffer(array.get_obj()).reshape(shape).astype(np.float32)
+
+        # get the computed Z-Index data as an array of float32 values
+        array = _global_shared_arrays[_KEY_RESULT_ZINDEX][_KEY_ARRAY]
+        shape = _global_shared_arrays[_KEY_RESULT_ZINDEX][_KEY_SHAPE]
+        zindex = np.frombuffer(array.get_obj()).reshape(shape).astype(np.float32)
+
+        # create a new variable to contain the SCPDSI values, assign into the dataset
+        long_name = "Self-calibrated Palmer Drought Severity Index"
+        scpdsi_attrs = {"long_name": long_name, "valid_min": -10.0, "valid_max": 10.0}
+        var_name_scpdsi = "scpdsi"
+        scpdsi_var = xr.Variable(dims=output_dims, data=scpdsi, attrs=scpdsi_attrs)
+        dataset[var_name_scpdsi] = scpdsi_var
+
+        # remove all data variables except for the new SCPDSI variable
+        for var_name in dataset.data_vars:
+            if var_name != var_name_scpdsi:
+                dataset = dataset.drop(var_name)
+
+        # TODO set global attributes accordingly for this new dataset
+
+        # write the dataset as NetCDF
+        netcdf_file_name = kwrgs["output_file_base"] + "_" + var_name_scpdsi + ".nc"
+        dataset.to_netcdf(netcdf_file_name)
+
+        # create a new variable to contain the PDSI values, assign into the dataset
+        long_name = "Palmer Drought Severity Index"
+        pdsi_attrs = {"long_name": long_name, "valid_min": -10.0, "valid_max": 10.0}
+        var_name_pdsi = "pdsi"
+        pdsi_var = xr.Variable(dims=output_dims, data=pdsi, attrs=pdsi_attrs)
+        dataset[var_name_pdsi] = pdsi_var
+
+        # remove all data variables except for the new PDSI variable
+        for var_name in dataset.data_vars:
+            if var_name != var_name_pdsi:
+                dataset = dataset.drop(var_name)
+
+        # TODO set global attributes accordingly for this new dataset
+
+        # write the dataset as NetCDF
+        netcdf_file_name = kwrgs["output_file_base"] + "_" + var_name_pdsi + ".nc"
+        dataset.to_netcdf(netcdf_file_name)
+
+        # create a new variable to contain the PHDI values, assign into the dataset
+        long_name = "Palmer Hydrological Drought Index"
+        phdi_attrs = {"long_name": long_name, "valid_min": -10.0, "valid_max": 10.0}
+        var_name_phdi = "phdi"
+        phdi_var = xr.Variable(dims=output_dims, data=phdi, attrs=phdi_attrs)
+        dataset[var_name_phdi] = phdi_var
+
+        # remove all data variables except for the new PHDI variable
+        for var_name in dataset.data_vars:
+            if var_name != var_name_phdi:
+                dataset = dataset.drop(var_name)
+
+        # TODO set global attributes accordingly for this new dataset
+
+        # write the dataset as NetCDF
+        netcdf_file_name = kwrgs["output_file_base"] + "_" + var_name_phdi + ".nc"
+        dataset.to_netcdf(netcdf_file_name)
+
+        # create a new variable to contain the PMDI values, assign into the dataset
+        long_name = "Palmer Modified Drought Index"
+        pmdi_attrs = {"long_name": long_name, "valid_min": -10.0, "valid_max": 10.0}
+        var_name_pmdi = "pmdi"
+        pmdi_var = xr.Variable(dims=output_dims, data=pmdi, attrs=pmdi_attrs)
+        dataset[var_name_pmdi] = pmdi_var
+
+        # remove all data variables except for the new PMDI variable
+        for var_name in dataset.data_vars:
+            if var_name != var_name_pmdi:
+                dataset = dataset.drop(var_name)
+
+        # TODO set global attributes accordingly for this new dataset
+
+        # write the dataset as NetCDF
+        netcdf_file_name = kwrgs["output_file_base"] + "_" + var_name_pmdi + ".nc"
+        dataset.to_netcdf(netcdf_file_name)
+
+        # create a new variable to contain the Z-Index values, assign into the dataset
+        long_name = "Palmer Z-Index"
+        zindex_attrs = {"long_name": long_name, "valid_min": -10.0, "valid_max": 10.0}
+        var_name_zindex = "zindex"
+        zindex_var = xr.Variable(dims=output_dims, data=zindex, attrs=zindex_attrs)
+        dataset[var_name_zindex] = zindex_var
+
+        # remove all data variables except for the new Z-Index variable
+        for var_name in dataset.data_vars:
+            if var_name != var_name_zindex:
+                dataset = dataset.drop(var_name)
+
+        # TODO set global attributes accordingly for this new dataset
+
+        # write the dataset as NetCDF
+        netcdf_file_name = kwrgs["output_file_base"] + "_" + var_name_zindex + ".nc"
+        dataset.to_netcdf(netcdf_file_name)
+
+    else:
+
+        # add an array to hold results to the dictionary of arrays
+        if _KEY_RESULT not in _global_shared_arrays:
+            _global_shared_arrays[_KEY_RESULT] = {
+                _KEY_ARRAY: multiprocessing.Array("d", int(np.prod(output_shape))),
+                _KEY_SHAPE: output_shape,
+            }
+
+        if keyword_arguments["index"] in ["spi", "pnp"]:
+
+            # apply the SPI function along the time axis (axis=2)
+            _parallel_process(
+                keyword_arguments["index"],
+                _global_shared_arrays,
+                {"var_name_precip": keyword_arguments["var_name_precip"]},
+                _KEY_RESULT,
+                args,
+            )
+
+        elif keyword_arguments["index"] == "spei":
+
+            # apply the SPEI function along the time axis (axis=2)
+            _parallel_process(
+                keyword_arguments["index"],
+                _global_shared_arrays,
+                {
+                    "var_name_precip": keyword_arguments["var_name_precip"],
+                    "var_name_pet": keyword_arguments["var_name_pet"],
+                },
+                _KEY_RESULT,
+                args,
+            )
+
+        elif keyword_arguments["index"] == "pet":
+
+            # create a shared memory array, wrap it as a numpy array and copy
+            # copy the data (values) from this variable's DataArray
+            da_lat = dataset["lat"]
+            shared_array = multiprocessing.Array("d", int(np.prod(da_lat.shape)))
+            shared_array_np = np.frombuffer(shared_array.get_obj()).reshape(
+                da_lat.shape
+            )
+            np.copyto(shared_array_np, da_lat.values)
+
+            # add to the dictionary of arrays
+            _global_shared_arrays[_KEY_LAT] = {
+                _KEY_ARRAY: shared_array,
+                _KEY_SHAPE: da_lat.shape,
+            }
+
+            # apply the PET function along the time axis (axis=2)
+            _parallel_process(
+                keyword_arguments["index"],
+                _global_shared_arrays,
+                {
+                    "var_name_temp": keyword_arguments["var_name_temp"],
+                    "var_name_lat": _KEY_LAT,
+                },
+                _KEY_RESULT,
+                args,
+            )
+
+        else:
+            raise ValueError(
+                "Unsupported index: '{index}'".format(index=keyword_arguments["index"])
+            )
+
+        # get the name and attributes to use for the index variable in the output NetCDF
+        output_var_name, output_var_attributes = _get_variable_attributes(
+            keyword_arguments
+        )
+
+        # get the shared memory results array and convert it to a numpy array
+        array = _global_shared_arrays[_KEY_RESULT][_KEY_ARRAY]
+        shape = _global_shared_arrays[_KEY_RESULT][_KEY_SHAPE]
+        index_values = np.frombuffer(array.get_obj()).reshape(shape).astype(np.float32)
+
+        # create a new variable to contain the index values, assign into the dataset
+        variable = xr.Variable(
+            dims=output_dims, data=index_values, attrs=output_var_attributes
+        )
+        dataset[output_var_name] = variable
+
+        # TODO set global attributes accordingly for this new dataset
+
+        # remove all data variables except for the new variable
+        for var_name in dataset.data_vars:
+            if var_name != output_var_name:
+                dataset = dataset.drop(var_name)
+
+        # write the dataset as NetCDF
+        netcdf_file_name = (
+            keyword_arguments["output_file_base"] + "_" + output_var_name + ".nc"
+        )
+        dataset.to_netcdf(netcdf_file_name)
+
+        return netcdf_file_name, output_var_name
+
+
+# ----------------------------------------------------------------------------------------------------------------------
+def _pet(temperatures, latitude, parameters):
+
+    return indices.pet(
+        temperature_celsius=temperatures,
+        latitude_degrees=latitude,
+        data_start_year=parameters["data_start_year"],
     )
 
-    # instantiate and run a grid processor object
-    grid_processor = GridProcessor(
-        index,
-        periodicity,
-        output_file_base,
-        calibration_start_year,
-        calibration_end_year,
-        netcdf_precip,
-        var_name_precip,
-        netcdf_temp,
-        var_name_temp,
-        netcdf_pet,
-        var_name_pet,
-        netcdf_awc,
-        var_name_awc,
-        scales,
+
+# ----------------------------------------------------------------------------------------------------------------------
+def _spi(precips, parameters):
+
+    return indices.spi(
+        values=precips,
+        scale=parameters["scale"],
+        distribution=parameters["distribution"],
+        data_start_year=parameters["data_start_year"],
+        calibration_year_initial=parameters["calibration_year_initial"],
+        calibration_year_final=parameters["calibration_year_final"],
+        periodicity=parameters["periodicity"],
     )
-    grid_processor.run()
+
+
+# ----------------------------------------------------------------------------------------------------------------------
+def _spei(precips, pet_mm, parameters):
+
+    return indices.spei(
+        precips_mm=precips,
+        pet_mm=pet_mm,
+        scale=parameters["scale"],
+        distribution=parameters["distribution"],
+        data_start_year=parameters["data_start_year"],
+        calibration_year_initial=parameters["calibration_year_initial"],
+        calibration_year_final=parameters["calibration_year_final"],
+        periodicity=parameters["periodicity"],
+    )
+
+
+# ----------------------------------------------------------------------------------------------------------------------
+def _palmers(precips, pet_mm, awc, parameters):
+
+    return indices.scpdsi(
+        precip_time_series=precips,
+        pet_time_series=pet_mm,
+        awc=awc,
+        data_start_year=parameters["data_start_year"],
+        calibration_start_year=parameters["calibration_start_year"],
+        calibration_end_year=parameters["calibration_end_year"],
+    )
+
+
+# ----------------------------------------------------------------------------------------------------------------------
+def _pnp(precips, parameters):
+
+    return indices.percentage_of_normal(
+        precips,
+        scale=parameters["scale"],
+        data_start_year=parameters["data_start_year"],
+        calibration_start_year=parameters["calibration_start_year"],
+        calibration_end_year=parameters["calibration_end_year"],
+        periodicity=parameters["periodicity"],
+    )
+
+
+# ----------------------------------------------------------------------------------------------------------------------
+def _init_worker(shared_arrays_dict):
+
+    global _global_shared_arrays
+    _global_shared_arrays = shared_arrays_dict
+
+
+# ----------------------------------------------------------------------------------------------------------------------
+def _parallel_process(index, arrays_dict, input_var_names, output_var_name, args):
+    """
+    Like numpy.apply_along_axis(), but takes advantage of multiple cores.
+
+    :param index:
+    :param arrays_dict:
+    :param input_var_names:
+    :param output_var_name:
+    :param args:
+    :return:
+    """
+
+    # find the start index of each sub-array we'll split out per worker process,
+    # assuming the shape of the output array is the same as all input arrays
+    shape = arrays_dict[output_var_name][_KEY_SHAPE]
+    d, m = divmod(shape[0], _NUMBER_OF_WORKER_PROCESSES)
+    split_indices = list(range(0, ((d + 1) * (m + 1)), (d + 1)))
+    if d != 0:
+        split_indices += list(range(split_indices[-1] + d, shape[0], d))
+
+    # build a list of parameters for each application of the function to an array chunk
+    chunk_params = []
+    if index in ["spi", "pnp"]:
+
+        if index == "spi":
+            func1d = _spi
+        else:
+            func1d = _pnp
+
+        # we have a single input array
+        for i in range(_NUMBER_OF_WORKER_PROCESSES):
+            params = {
+                "index": index,
+                "func1d": func1d,
+                "input_var_name": input_var_names["var_name_precip"],
+                "output_var_name": output_var_name,
+                "sub_array_start": split_indices[i],
+                "args": args,
+            }
+            if i < (_NUMBER_OF_WORKER_PROCESSES - 1):
+                params["sub_array_end"] = split_indices[i + 1]
+            else:
+                params["sub_array_end"] = None
+
+            chunk_params.append(params)
+
+    elif index == "spei":
+
+        for i in range(_NUMBER_OF_WORKER_PROCESSES):
+            params = {
+                "index": index,
+                "func1d": _spei,
+                "var_name_precip": input_var_names["var_name_precip"],
+                "var_name_pet": input_var_names["var_name_pet"],
+                "output_var_name": output_var_name,
+                "sub_array_start": split_indices[i],
+                "args": args,
+            }
+            if i < (_NUMBER_OF_WORKER_PROCESSES - 1):
+                params["sub_array_end"] = split_indices[i + 1]
+            else:
+                params["sub_array_end"] = None
+
+            chunk_params.append(params)
+
+    elif index == "pet":
+
+        for i in range(_NUMBER_OF_WORKER_PROCESSES):
+            params = {
+                "index": index,
+                "func1d": _pet,
+                "var_name_temp": input_var_names["var_name_temp"],
+                "var_name_lat": input_var_names["var_name_lat"],
+                "output_var_name": output_var_name,
+                "sub_array_start": split_indices[i],
+                "args": args,
+            }
+            if i < (_NUMBER_OF_WORKER_PROCESSES - 1):
+                params["sub_array_end"] = split_indices[i + 1]
+            else:
+                params["sub_array_end"] = None
+
+            chunk_params.append(params)
+
+    elif index == "palmers":
+
+        for i in range(_NUMBER_OF_WORKER_PROCESSES):
+            params = {
+                "index": index,
+                "func1d": _palmers,
+                "var_name_precip": input_var_names["var_name_precip"],
+                "var_name_pet": input_var_names["var_name_pet"],
+                "var_name_awc": input_var_names["var_name_awc"],
+                "output_var_name": output_var_name,
+                "sub_array_start": split_indices[i],
+                "args": args,
+            }
+            if i < (_NUMBER_OF_WORKER_PROCESSES - 1):
+                params["sub_array_end"] = split_indices[i + 1]
+            else:
+                params["sub_array_end"] = None
+
+            chunk_params.append(params)
+
+    else:
+        raise ValueError("Unsupported index: {index}".format(index=index))
+
+    # instantiate a process pool
+    with multiprocessing.Pool(
+        processes=_NUMBER_OF_WORKER_PROCESSES,
+        initializer=_init_worker,
+        initargs=(arrays_dict,),
+    ) as pool:
+
+        """
+         the function _unpacking_apply_along_axis() being applied in Pool.map() is separate
+         so that subprocesses can import it, and is simply a thin wrapper that handles the
+         fact that Pool.map() only takes a single argument:
+        """
+
+        if index in ["spei", "pet"]:
+            pool.map(_unpacking_apply_along_axis_double, chunk_params)
+        elif index == "palmers":
+            pool.map(_unpacking_apply_along_axis_palmers, chunk_params)
+        else:
+            pool.map(_unpacking_apply_along_axis, chunk_params)
+
+
+# ----------------------------------------------------------------------------------------------------------------------
+def _unpacking_apply_along_axis(params):
+    """
+    Like numpy.apply_along_axis(), but and with arguments in a dict
+    instead.
+
+    This function is useful with multiprocessing.Pool().map(): (1)
+    map() only handles functions that take a single argument, and (2)
+    this function can generally be imported from a module, as required
+    by map().
+    """
+    func1d = params["func1d"]
+    start_index = params["sub_array_start"]
+    end_index = params["sub_array_end"]
+    array = _global_shared_arrays[params["input_var_name"]][_KEY_ARRAY]
+    shape = _global_shared_arrays[params["input_var_name"]][_KEY_SHAPE]
+    np_array = np.frombuffer(array.get_obj()).reshape(shape)
+    sub_array = np_array[start_index:end_index]
+    args = params["args"]
+
+    computed_array = np.apply_along_axis(func1d, axis=2, arr=sub_array, parameters=args)
+
+    output_array = _global_shared_arrays[params["output_var_name"]][_KEY_ARRAY]
+    np_output_array = np.frombuffer(output_array.get_obj()).reshape(shape)
+    np.copyto(np_output_array[start_index:end_index], computed_array)
+
+
+# ----------------------------------------------------------------------------------------------------------------------
+def _unpacking_apply_along_axis_double(params):
+    """
+    Like numpy.apply_along_axis(), but and with arguments in a dict
+    instead.
+
+    This function is useful with multiprocessing.Pool().map(): (1)
+    map() only handles functions that take a single argument, and (2)
+    this function can generally be imported from a module, as required
+    by map().
+    """
+
+    func1d = params["func1d"]
+    start_index = params["sub_array_start"]
+    end_index = params["sub_array_end"]
+    if params["index"] == "pet":
+        first_array_key = params["var_name_temp"]
+        second_array_key = params["var_name_lat"]
+    elif params["index"] == "spei":
+        first_array_key = params["var_name_precip"]
+        second_array_key = params["var_name_pet"]
+    else:
+        raise ValueError("Unsupported index: {index}".format(index=params["index"]))
+
+    shape = _global_shared_arrays[params["output_var_name"]][_KEY_SHAPE]
+    first_array = _global_shared_arrays[first_array_key][_KEY_ARRAY]
+    first_np_array = np.frombuffer(first_array.get_obj()).reshape(shape)
+    sub_array_1 = first_np_array[start_index:end_index]
+    if params["index"] == "pet":
+        second_array = _global_shared_arrays[second_array_key][_KEY_ARRAY]
+        second_np_array = np.frombuffer(second_array.get_obj()).reshape(shape[0])
+    else:
+        second_array = _global_shared_arrays[second_array_key][_KEY_ARRAY]
+        second_np_array = np.frombuffer(second_array.get_obj()).reshape(shape)
+    sub_array_2 = second_np_array[start_index:end_index]
+
+    # get the output shared memory array, convert to numpy, and get the subarray slice
+    output_array = _global_shared_arrays[params["output_var_name"]][_KEY_ARRAY]
+    computed_array = np.frombuffer(output_array.get_obj()).reshape(shape)[
+        start_index:end_index
+    ]
+
+    for i, (x, y) in enumerate(zip(sub_array_1, sub_array_2)):
+        for j in range(x.shape[0]):
+            if params["index"] == "pet":
+                computed_array[i, j] = func1d(x[j], y, parameters=params["args"])
+            else:
+                computed_array[i, j] = func1d(x[j], y[j], parameters=params["args"])
+
+
+# ----------------------------------------------------------------------------------------------------------------------
+def _unpacking_apply_along_axis_palmers(params):
+    """
+    Like numpy.apply_along_axis(), but and with arguments in a dict
+    instead.
+
+    This function is useful with multiprocessing.Pool().map(): (1)
+    map() only handles functions that take a single argument, and (2)
+    this function can generally be imported from a module, as required
+    by map().
+    """
+    func1d = params["func1d"]
+    start_index = params["sub_array_start"]
+    end_index = params["sub_array_end"]
+    precip_array_key = params["var_name_precip"]
+    pet_array_key = params["var_name_pet"]
+    awc_array_key = params["var_name_awc"]
+
+    shape = _global_shared_arrays[params["output_var_name"]][_KEY_SHAPE]
+    precip_array = _global_shared_arrays[precip_array_key][_KEY_ARRAY]
+    precip_np_array = np.frombuffer(precip_array.get_obj()).reshape(shape)
+    sub_array_precip = precip_np_array[start_index:end_index]
+    pet_array = _global_shared_arrays[pet_array_key][_KEY_ARRAY]
+    pet_np_array = np.frombuffer(pet_array.get_obj()).reshape(shape)
+    sub_array_pet = pet_np_array[start_index:end_index]
+    awc_array = _global_shared_arrays[awc_array_key][_KEY_ARRAY]
+    awc_np_array = np.frombuffer(awc_array.get_obj()).reshape([shape[0], shape[1]])
+    sub_array_awc = awc_np_array[start_index:end_index]
+
+    args = params["args"]
+
+    # get the output shared memory arrays, convert to numpy, and get the subarray slices
+    scpdsi_output_array = _global_shared_arrays[_KEY_RESULT_SCPDSI][_KEY_ARRAY]
+    scpdsi = np.frombuffer(scpdsi_output_array.get_obj()).reshape(shape)[
+        start_index:end_index
+    ]
+
+    pdsi_output_array = _global_shared_arrays[_KEY_RESULT_PDSI][_KEY_ARRAY]
+    pdsi = np.frombuffer(pdsi_output_array.get_obj()).reshape(shape)[
+        start_index:end_index
+    ]
+
+    phdi_output_array = _global_shared_arrays[_KEY_RESULT_PHDI][_KEY_ARRAY]
+    phdi = np.frombuffer(phdi_output_array.get_obj()).reshape(shape)[
+        start_index:end_index
+    ]
+
+    pmdi_output_array = _global_shared_arrays[_KEY_RESULT_PMDI][_KEY_ARRAY]
+    pmdi = np.frombuffer(pmdi_output_array.get_obj()).reshape(shape)[
+        start_index:end_index
+    ]
+
+    zindex_output_array = _global_shared_arrays[_KEY_RESULT_ZINDEX][_KEY_ARRAY]
+    zindex = np.frombuffer(zindex_output_array.get_obj()).reshape(shape)[
+        start_index:end_index
+    ]
+
+    for i, (precip, pet, awc) in enumerate(
+        zip(sub_array_precip, sub_array_pet, sub_array_awc)
+    ):
+        for j in range(precip.shape[0]):
+            scpdsi[i, j], pdsi[i, j], phdi[i, j], pmdi[i, j], zindex[i, j] = func1d(
+                precip[j], pet[j], awc[j], parameters=args
+            )
+
+
+# ----------------------------------------------------------------------------------------------------------------------
+def _prepare_file(netcdf_file, var_name):
+    """
+    Determine if the NetCDF file has the expected lat, lon, and time dimensions, and if not
+    correctly ordered then create a temporary NetCDF with dimensions in (lat, lon, time) order,
+    otherwise just return the input NetCDF unchanged.
+
+    :param str netcdf_file:
+    :param str var_name:
+    :return:
+    """
+
+    # make sure we have lat, lon, and time as variable dimensions, regardless of order
+    ds = xr.open_dataset(netcdf_file)
+    if len(ds[var_name].dims) == 2:
+        expected_dims = ("lat", "lon")
+        dims = "lat,lon"
+    elif len(ds[var_name].dims) == 3:
+        expected_dims = ("lat", "lon", "time")
+        dims = "lat,lon,time"
+    else:
+        raise ValueError(
+            "Unsupported dimensions for variable '{var_name}': {dims}".format(
+                var_name=var_name, dims=ds[var_name].dims
+            )
+        )
+
+    if Counter(ds[var_name].dims) != Counter(expected_dims):
+        message = "Invalid dimensions for variable '{var_name}': {dims}".format(
+            var_name=var_name, dims=ds[var_name].dims
+        )
+        _logger.error(message)
+        raise ValueError(message)
+
+    # perform reorder of dimensions if necessary
+    if ds[var_name].dims != expected_dims:
+        nco = Nco()
+        netcdf_file = nco.ncpdq(
+            input=netcdf_file, options=['-a \\"{dims}\\"'.format(dims=dims), "-O"]
+        )
+
+    return netcdf_file
 
 
 # ----------------------------------------------------------------------------------------------------------------------
 if __name__ == "__main__":
     """
-    This module is used to perform climate indices processing on gridded datasets in NetCDF.
+    This script is used to perform climate indices processing on gridded datasets in NetCDF.
 
     Example command line arguments for SPI only using monthly precipitation input:
-    
-    --index spi 
+
+    --index spi
     --periodicity monthly
-    --scales 1 2 3 6 9 12 24 
-    --calibration_start_year 1998 
-    --calibration_end_year 2016 
-    --netcdf_precip example_data/nclimgrid_prcp_lowres.nc 
-    --var_name_precip prcp 
-    --output_file_base ~/data/test/nclimgrid_lowres 
+    --scales 1 2 3 6 9 12 24
+    --calibration_start_year 1998
+    --calibration_end_year 2016
+    --netcdf_precip example_data/nclimgrid_prcp_lowres.nc
+    --var_name_precip prcp
+    --output_file_base ~/data/test/spi/nclimgrid_lowres
     """
 
     try:
@@ -1290,7 +1257,7 @@ if __name__ == "__main__":
         parser.add_argument(
             "--index",
             help="Indices to compute",
-            choices=["spi", "spei", "pnp", "scaled", "pet", "palmers"],
+            choices=["spi", "spei", "pnp", "scaled", "pet", "palmers", "all"],
             required=True,
         )
         parser.add_argument(
@@ -1350,24 +1317,183 @@ if __name__ == "__main__":
             help="Base output file path and name for the resulting output files",
             required=True,
         )
-        args = parser.parse_args()
-
-        process_grid(
-            args.index,
-            args.periodicity,
-            args.output_file_base,
-            args.calibration_start_year,
-            args.calibration_end_year,
-            args.netcdf_precip,
-            args.var_name_precip,
-            args.netcdf_temp,
-            args.var_name_temp,
-            args.netcdf_pet,
-            args.var_name_pet,
-            args.netcdf_awc,
-            args.var_name_awc,
-            args.scales,
+        parser.add_argument(
+            "--multiprocessing",
+            help="Indices to compute",
+            choices=["single", "all_but_one", "all"],
+            required=False,
+            default="all_but_one",
         )
+        arguments = parser.parse_args()
+
+        # validate the arguments
+        _validate_args(arguments)
+
+        if arguments.multiprocessing == "single":
+            _NUMBER_OF_WORKER_PROCESSES = 1
+        elif arguments.multiprocessing == "all":
+            _NUMBER_OF_WORKER_PROCESSES = multiprocessing.cpu_count()
+        else:  # default ("all_but_one")
+            _NUMBER_OF_WORKER_PROCESSES = multiprocessing.cpu_count() - 1
+
+        # compute SPI if specified
+        if arguments.index in ["spi", "scaled", "all"]:
+
+            # prepare precipitation NetCDF in case dimensions not (lat, lon, time) or if any coordinates are descending
+            netcdf_precip = _prepare_file(
+                arguments.netcdf_precip, arguments.var_name_precip
+            )
+
+            # run SPI computations for each scale/distribution in turn
+            for scale in arguments.scales:
+                for dist in indices.Distribution:
+
+                    # keyword arguments used for the SPI function
+                    kwrgs = {
+                        "index": "spi",
+                        "netcdf_precip": netcdf_precip,
+                        "var_name_precip": arguments.var_name_precip,
+                        "scale": scale,
+                        "distribution": dist,
+                        "periodicity": arguments.periodicity,
+                        "calibration_start_year": arguments.calibration_start_year,
+                        "calibration_end_year": arguments.calibration_end_year,
+                        "output_file_base": arguments.output_file_base,
+                    }
+
+                    # compute and write SPI
+                    _compute_write_index(kwrgs)
+
+            # remove temporary file if one was created
+            if netcdf_precip != arguments.netcdf_precip:
+                os.remove(netcdf_precip)
+
+        if arguments.index in ["pet", "spei", "scaled", "palmers", "all"]:
+
+            # run PET computation only if we've not been provided with a PET file
+            if arguments.netcdf_pet is None:
+
+                # prepare temperature NetCDF in case dimensions not (lat, lon, time)
+                # or if coordinates are descending
+                netcdf_temp = _prepare_file(
+                    arguments.netcdf_temp, arguments.var_name_temp
+                )
+
+                # keyword arguments used for the PET function
+                kwargs = {
+                    "index": "pet",
+                    "netcdf_temp": netcdf_temp,
+                    "var_name_temp": arguments.var_name_temp,
+                    "output_file_base": arguments.output_file_base,
+                }
+
+                # run PET computation, getting the PET file and corresponding variable name for later use
+                arguments.netcdf_pet, arguments.var_name_pet = _compute_write_index(
+                    kwargs
+                )
+
+                # remove temporary file
+                if netcdf_temp != arguments.netcdf_temp:
+                    os.remove(netcdf_temp)
+
+        if arguments.index in ["spei", "scaled", "all"]:
+
+            # prepare NetCDFs in case dimensions not (lat, lon, time) or if any coordinates are descending
+            netcdf_precip = _prepare_file(
+                arguments.netcdf_precip, arguments.var_name_precip
+            )
+            netcdf_pet = _prepare_file(arguments.netcdf_pet, arguments.var_name_pet)
+
+            # run SPEI computations for each scale/distribution in turn
+            for scale in arguments.scales:
+                for dist in indices.Distribution:
+
+                    # keyword arguments used for the SPI function
+                    kwrgs = {
+                        "index": "spei",
+                        "netcdf_precip": netcdf_precip,
+                        "var_name_precip": arguments.var_name_precip,
+                        "netcdf_pet": netcdf_pet,
+                        "var_name_pet": arguments.var_name_pet,
+                        "scale": scale,
+                        "distribution": dist,
+                        "periodicity": arguments.periodicity,
+                        "calibration_start_year": arguments.calibration_start_year,
+                        "calibration_end_year": arguments.calibration_end_year,
+                        "output_file_base": arguments.output_file_base,
+                    }
+
+                    # compute and write SPEI
+                    _compute_write_index(kwrgs)
+
+            # remove temporary file if one was created
+            if netcdf_precip != arguments.netcdf_precip:
+                os.remove(netcdf_precip)
+            if netcdf_pet != arguments.netcdf_pet:
+                os.remove(netcdf_pet)
+
+        if arguments.index in ["pnp", "scaled", "all"]:
+
+            # prepare precipitation NetCDF in case dimensions not (lat, lon, time) or if any coordinates are descending
+            netcdf_precip = _prepare_file(
+                arguments.netcdf_precip, arguments.var_name_precip
+            )
+
+            # run PNP computations for each scale in turn
+            for scale in arguments.scales:
+
+                # keyword arguments used for the SPI function
+                kwrgs = {
+                    "index": "pnp",
+                    "netcdf_precip": netcdf_precip,
+                    "var_name_precip": arguments.var_name_precip,
+                    "scale": scale,
+                    "periodicity": arguments.periodicity,
+                    "calibration_start_year": arguments.calibration_start_year,
+                    "calibration_end_year": arguments.calibration_end_year,
+                    "output_file_base": arguments.output_file_base,
+                }
+
+                # compute and write PNP
+                _compute_write_index(kwrgs)
+
+            # remove temporary precipitation file if one was created
+            if netcdf_precip != arguments.netcdf_precip:
+                os.remove(netcdf_precip)
+
+        if arguments.index in ["palmers", "all"]:
+
+            # prepare NetCDFs in case dimensions not (lat, lon, time)
+            netcdf_precip = _prepare_file(
+                arguments.netcdf_precip, arguments.var_name_precip
+            )
+            netcdf_pet = _prepare_file(arguments.netcdf_pet, arguments.var_name_pet)
+            netcdf_awc = _prepare_file(arguments.netcdf_awc, arguments.var_name_awc)
+
+            # keyword arguments used for the SPI function
+            kwrgs = {
+                "index": "palmers",
+                "netcdf_precip": netcdf_precip,
+                "var_name_precip": arguments.var_name_precip,
+                "netcdf_pet": netcdf_pet,
+                "var_name_pet": arguments.var_name_pet,
+                "netcdf_awc": netcdf_awc,
+                "var_name_awc": arguments.var_name_awc,
+                "calibration_start_year": arguments.calibration_start_year,
+                "calibration_end_year": arguments.calibration_end_year,
+                "output_file_base": arguments.output_file_base,
+            }
+
+            # compute and write Palmers
+            _compute_write_index(kwrgs)
+
+            # remove temporary files if they were created
+            if netcdf_precip != arguments.netcdf_precip:
+                os.remove(netcdf_precip)
+            if netcdf_pet != arguments.netcdf_pet:
+                os.remove(netcdf_pet)
+            if netcdf_awc != arguments.netcdf_awc:
+                os.remove(netcdf_awc)
 
         # report on the elapsed time
         end_datetime = datetime.now()
